Refactor SCRJump import and add tests for SCRJump functionality
