+++ src/dycov/gfm/calculators/calculator_factory.py
@@ -14,7 +14,7 @@ from dycov.gfm.calculators.amplitude_step import AmplitudeStep
 from dycov.gfm.calculators.gfm_calculator import GFMCalculator
 from dycov.gfm.calculators.phase_jump import PhaseJump
 from dycov.gfm.calculators.rocof import RoCoF
-from dycov.gfm.calculators.scrjump import SCRJump
+from dycov.gfm.calculators.scr_jump import SCRJump
 from dycov.gfm.parameters import GFMParameters
