+++ src/dycov/templates/PCS/gfm/PCS_RTE-IGFM3/PCSDescription.ini
@@ -55,21 +55,25 @@ P0 = 0
 Q0 = 0
 U0 = 1
 SavedBattery = 0.5
+
 [PCS_RTE-IGFM3.S_SCRup1.OC2]
 P0 = 0.5*Pmax
 Q0 = 0
 U0 = 1
 SavedBattery = 0.5
+
 [PCS_RTE-IGFM3.S_SCRup1.OC3]
 P0 = -0.5*Pmax
 Q0 = 0
 U0 = 1
 SavedBattery = 0.5
+
 [PCS_RTE-IGFM3.S_SCRup1.OC6]
 P0 = 0
 Q0 = 0
 U0 = 1
 SavedBattery = 0.5
+
 [PCS_RTE-IGFM3.S_SCRup1.OC7]
 P0 = 0
 Q0 = 0
@@ -95,21 +99,25 @@ P0 = 0
 Q0 = 0
 U0 = 1
 SavedBattery = 0.5
+
 [PCS_RTE-IGFM3.S_SCRdown1.OC2]
 P0 = 0.5*Pmax
 Q0 = 0
 U0 = 1
 SavedBattery = 0.5
+
 [PCS_RTE-IGFM3.S_SCRdown1.OC3]
 P0 = -0.5*Pmax
 Q0 = 0
 U0 = 1
 SavedBattery = 0.5
+
 [PCS_RTE-IGFM3.S_SCRdown1.OC6]
 P0 = 0
 Q0 = 0
 U0 = 1
 SavedBattery = 0.5
+
 [PCS_RTE-IGFM3.S_SCRdown1.OC7]
 P0 = 0
 Q0 = 0
@@ -135,21 +143,25 @@ P0 = 0
 Q0 = 0
 U0 = 1
 SavedBattery = 0.5
+
 [PCS_RTE-IGFM3.S_SCRup2.OC2]
 P0 = 0.5*Pmax
 Q0 = 0
 U0 = 1
 SavedBattery = 0.5
+
 [PCS_RTE-IGFM3.S_SCRup2.OC3]
 P0 = -0.5*Pmax
 Q0 = 0
 U0 = 1
 SavedBattery = 0.5
+
 [PCS_RTE-IGFM3.S_SCRup2.OC6]
 P0 = 0
 Q0 = 0
 U0 = 1
 SavedBattery = 0.5
+
 [PCS_RTE-IGFM3.S_SCRup2.OC7]
 P0 = 0
 Q0 = 0
@@ -175,21 +187,25 @@ P0 = 0
 Q0 = 0
 U0 = 1
 SavedBattery = 0.5
+
 [PCS_RTE-IGFM3.S_SCRdown2.OC2]
 P0 = 0.5*Pmax
 Q0 = 0
 U0 = 1
 SavedBattery = 0.5
+
 [PCS_RTE-IGFM3.S_SCRdown2.OC3]
 P0 = -0.5*Pmax
 Q0 = 0
 U0 = 1
 SavedBattery = 0.5
+
 [PCS_RTE-IGFM3.S_SCRdown2.OC6]
 P0 = 0
 Q0 = 0
 U0 = 1
 SavedBattery = 0.5
+
 [PCS_RTE-IGFM3.S_SCRdown2.OC7]
 P0 = 0
 Q0 = 0
